Test: move Gradle daemon memory test to separate class
